Reordering drag and drop dynamically
As completing drag and drop using DOM manipulation causes issue with
Angular i.e. the content object updates successfully, but the collection
items appear in the wrong order (as they would have been moved around
within the DOM). By completing the reordering dynamically, ensuring that
the correct content is reflected immediatley. Note - if Angular provided
a method to force a components children to be removed and redrawn, this
using DOM manipulation would not be an issue