Creating basis of WYSIWYG HTML editor
Emitting events through the collection-item/collection/content-editor
encapsuled components, so as to deal with file upload, view and
selecting of images from the media items gallery using the existing
functions. Passing callback functions as the event payload, so that the
default behaviour of the content-editor (which is to update the content
for the item that selected the file) to instead return the url of the
file to the WYSIWYG editor. Storing the returned url on a static
property of the WysiwygHtmlComponent, as when the method is invoked by
the ContentEditorComponent, the "this" keyword would refer to that
component instead of the wysiwyg one.