Fixes to code editor - "reset" button and updating from structure
Previously the reset button would only work once - now works
indefinatley. Also, since the changes to validating and sanitising the
structure, the code was not refreshing when the project structure was
returned from the server (similar issue as with arrays not updating)
calling split() to trick Angular into thinking it is a new object.