Fix to "prevChar" not working on first backspace in code editor
The only way to force Angular to update the content of the textarea when
it is refreshed from the server is to "split" the string. Forgot to
rejoin it once it gets to the code editor! (Normally not an issue, as
adding a new character was parsing it to a string anyway, but at least
now backspace never throws an error)